On task click set modal contents to task contents
When a task item is clicked, open the modal
Then set the title and description of the modal
to the name and description of
the task

On modal close, clear the modal and close

On modal save, print the contents of modal title and description to console.
It will be the same as the task that called it.